Comparison of decomposition methods
Implemented a few lines to check correlation between a few mesh rows (top row, middle row and bottom row). However, it was only a few points to compare so the analysis may not be reliable.
Thus, a basic script was implemented with a small covariance matrix and a lot of points. All 3 decomposition methods worked: svd, cholesky and eigen. This script is saved as 'CorrelatedData.m'